requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
pyyaml==6.0
joblib==1.3.1
python-dateutil==2.8.2
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
streamlit==1.31.1
folium==0.14.0
streamlit-folium==0.17.0
plotly==5.18.0
